Refactor and split up databricks node actions
