+++ packages/nodes-base/nodes/Databricks/actions/databricksSql/DatabricksSql.resource.ts
@@ -1,0 +1,1 @@
+export * as executeQuery from './executeQuery.operation';
