+++ packages/nodes-base/nodes/Databricks/actions/files/Files.resource.ts
@@ -1,0 +1,7 @@
+export * as uploadFile from './uploadFile.operation';
+export * as downloadFile from './downloadFile.operation';
+export * as deleteFile from './deleteFile.operation';
+export * as getFileInfo from './getFileInfo.operation';
+export * as listDirectory from './listDirectory.operation';
+export * as createDirectory from './createDirectory.operation';
+export * as deleteDirectory from './deleteDirectory.operation';
