+++ packages/nodes-base/nodes/Databricks/actions/genie/Genie.resource.ts
@@ -1,0 +1,6 @@
+export * as createMessage from './createMessage.operation';
+export * as getMessage from './getMessage.operation';
+export * as getQueryResults from './getQueryResults.operation';
+export * as executeMessageQuery from './executeMessageQuery.operation';
+export * as getSpace from './getSpace.operation';
+export * as startConversation from './startConversation.operation';
