+++ packages/nodes-base/nodes/Databricks/actions/genie/startConversation.operation.ts
@@ -1,0 +1,19 @@
+import type { IExecuteFunctions, INodeExecutionData } from 'n8n-workflow';
+
+import { getActiveCredentialType, getHost } from '../helpers';
+
+export async function execute(this: IExecuteFunctions, i: number): Promise<INodeExecutionData[]> {
+	const credentialType = getActiveCredentialType(this, i);
+	const host = await getHost(this, credentialType);
+	const spaceId = this.getNodeParameter('spaceId', i) as string;
+
+	const response = await this.helpers.httpRequestWithAuthentication.call(this, credentialType, {
+		method: 'POST',
+		url: `${host}/api/2.0/genie/spaces/${spaceId}/start-conversation`,
+		body: { content: this.getNodeParameter('initialMessage', i) as string },
+		headers: { 'Content-Type': 'application/json' },
+		json: true,
+	});
+
+	return [{ json: response, pairedItem: { item: i } }];
+}
