+++ packages/nodes-base/nodes/Databricks/actions/modelServing/ModelServing.resource.ts
@@ -1,0 +1,1 @@
+export * as queryEndpoint from './queryEndpoint.operation';
