+++ packages/nodes-base/nodes/Databricks/actions/vectorSearch/VectorSearch.resource.ts
@@ -1,0 +1,4 @@
+export * as queryIndex from './queryIndex.operation';
+export * as createIndex from './createIndex.operation';
+export * as getIndex from './getIndex.operation';
+export * as listIndexes from './listIndexes.operation';
